* develop: (22 commits)
  Mise à jour du composer
  Fix deprecated
  Update CachePathResolver.php
  Update imagine.xml
  Update README.md
  Update README.md
  Update README.md
  Add permissions configuration setting to set specific chmod value on generated images.
  Update composer.json
  can be overrided in filter options.
  fixed link to imagine library
  Fixing bug related to concurrency in dir creation. See PHP bug #35326.
  Fixing bug in cached content verifying (switching from 'file_exists' to 'is_file').
  Merge recursively the config
  Check if $cached is not a directory to avoid error on unlink
  Fixed bug with symfony 2.3
  added missing use statement
  Update composer.json
  catch exception when invalid filter is used
  Add docs for new crop filter
  ...

+++ Imagine/CachePathResolver.php
@@ -4,6 +4,7 @@ namespace Avalanche\Bundle\ImagineBundle\Imagine;
 
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\Routing\RouterInterface;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
 
 class CachePathResolver
 {
@@ -11,6 +12,11 @@ class CachePathResolver
      * @var string
      */
     private $webRoot;
+
+    /**
+     * @var string
+     */
+    private $sourceRoot;
 
     /**
      * @var Symfony\Component\Routing\RouterInterface
@@ -21,11 +27,13 @@ class CachePathResolver
      * Constructs cache path resolver with a given web root and cache prefix
      *
      * @param string                                    $webRoot
+     * @param string                                    $sourceRoot
      * @param Symfony\Component\Routing\RouterInterface $router
      */
-    public function __construct($webRoot, RouterInterface $router)
+    public function __construct($webRoot, $sourceRoot, RouterInterface $router)
     {
         $this->webRoot = $webRoot;
+        $this->sourceRoot = $sourceRoot;
         $this->router  = $router;
     }
 
@@ -36,14 +44,23 @@ class CachePathResolver
      * @param string $filter
      * @param boolean $absolute
      */
-    public function getBrowserPath($path, $filter, $absolute = false)
+    public function getBrowserPath($path, $filter, $absolute = UrlGeneratorInterface::RELATIVE_PATH)
     {
-        // identify if current path is not under specified web root and return
+        // identify if current path is not under specified source root and return
         // unmodified path in that case
-        $realPath = realpath($this->webRoot.$path);
+        $realPath = realpath($this->sourceRoot.$path);
 
-        if (!0 === strpos($realPath, $this->webRoot)) {
+        if (!0 === strpos($realPath, $this->sourceRoot)) {
             return $path;
+        }
+
+        switch ($absolute) {
+            case true:
+                $absolute = UrlGeneratorInterface::ABSOLUTE_URL;
+                break;
+            case false:
+                $absolute = UrlGeneratorInterface::RELATIVE_PATH;
+                break;
         }
 
         $path = str_replace(
@@ -53,10 +70,10 @@ class CachePathResolver
                 'path' => ltrim($path, '/')
             ), $absolute)
         );
-        
+
         $cached = realpath($this->webRoot.$path);
 
-        if (file_exists($cached) && filemtime($realPath) > filemtime($cached)) {
+        if (file_exists($cached) && !is_dir($cached) && filemtime($realPath) > filemtime($cached)) {
             unlink($cached);
         }
